+++ libticalcs/trunk/src/cmd73.c
@@ -120,7 +120,7 @@ TIEXPORT3 int TICALL ti73_send_CTS(CalcHandle* handle)
 	return dbus_send(handle, PC_TI7383, CMD_CTS, 0, NULL);
 }
 
-TIEXPORT3 int TICALL ti73_send_XDP(CalcHandle* handle, int length, uint8_t * data)
+TIEXPORT3 int TICALL ti73_send_XDP(CalcHandle* handle, uint16_t length, uint8_t * data)
 {
 	if (handle == NULL)
 	{
