+++ libticalcs/trunk/src/cmd82.c
@@ -90,7 +90,7 @@ TIEXPORT3 int TICALL ti82_send_CTS(CalcHandle* handle)
 	return dbus_send(handle, PC_TI8283, CMD_CTS, 0, NULL);
 }
 
-TIEXPORT3 int TICALL ti82_send_XDP(CalcHandle* handle, int length, uint8_t * data)
+TIEXPORT3 int TICALL ti82_send_XDP(CalcHandle* handle, uint16_t length, uint8_t * data)
 {
 	if (handle == NULL)
 	{
